refactor(ml): split ml_engine.py into modules with protocols and factory
